+++ src/components/Console/CommandLine.tsx
@@ -1,8 +1,14 @@
 import React, { FC } from 'react'
-import { COMMAND_CODE, commandLine } from '../../data/commands'
+import { COMMAND_CODE } from '../../data/commands'
 import styled from 'styled-components'
 
-const CommandLine : FC<commandLine> = ({infoType,command,info}) => {
+export interface ICommandLine {
+  command: string,
+  info?: string,
+  infoType: COMMAND_CODE
+}
+
+const CommandLine : FC<ICommandLine> = ({infoType,command,info}) => {
   return (
     <StyledCommandLine>
         {command && <StyledCommand>{command}</StyledCommand>}
@@ -26,4 +32,4 @@ const StyledCommandInfo = styled.p<{type: COMMAND_CODE}>`
     color: ${(props) => props.type === COMMAND_CODE.ERROR ? "var(--color-error)" : "white"};
 `;
 
-export default CommandLine
+export default React.memo(CommandLine)