+++ src/features/desktop/desktopSlice.tsx
@@ -2,9 +2,8 @@ import { createSlice } from "@reduxjs/toolkit";
 import type { PayloadAction } from "@reduxjs/toolkit";
 import { PANELS } from '../../data/panels';
 import { recipeKey } from '../../data/recipes';
-import { DISHES, DisplayedDish } from "../../data/dishes";
+import { DISHES, DisplayedDish, MAX_DISHES_ON_SCREEN } from "../../data/dishes";
 import { findNextId } from "../../helpers/array.helper";
-import React, { RefObject } from "react";
 
 type VisiblePanels = {[key in PANELS]?: boolean} 
 type VisibleDishes = {[key in DISHES]: Array<number>};
@@ -15,8 +14,6 @@ export type DesktopState = {
     dishes: VisibleDishes,
     recipes: VisibleRecipes,
 }
-
-const MAXED_DISHES_ON_SCREEN : number = 5;
 
 const initialState: DesktopState = {
     dishes: {
@@ -31,7 +28,7 @@ const initialState: DesktopState = {
         [PANELS.FRIDGE]: false,
         [PANELS.WORKTOP]: false,
         [PANELS.RECIPEBOOK]: false,
-    },
+    }
 };
 
 export const desktopSlice = createSlice({
@@ -39,7 +36,8 @@ export const desktopSlice = createSlice({
     initialState,
     reducers: {
         toggleOpenPanel: (state, action: PayloadAction<{panelType: PANELS, opened?: boolean}>) => {
-            state.panels[action.payload.panelType] = action.payload.opened;
+            const panels = { ...state.panels, [action.payload.panelType]: action.payload.opened };
+            state.panels = panels;
         },
         displayRecipe: (state, action: PayloadAction<{recipe: recipeKey,display:boolean}>) => {
             const name = action.payload.recipe;
@@ -52,14 +50,14 @@ export const desktopSlice = createSlice({
                 state.recipes.push(name);
             }
         },
-        displayDish: (state, action: PayloadAction<DISHES>) => {
+        displayDish: (state, action: PayloadAction<DISHES>)=> {
             const dishType = action.payload;
             const id = findNextId([...state.dishes[dishType]]);
-            if(state.dishes[dishType].length < MAXED_DISHES_ON_SCREEN){
+            if(state.dishes[dishType].length < MAX_DISHES_ON_SCREEN){
                 state.dishes[dishType].push(id);
             }
         },
-        removeDish: (state, action: PayloadAction<DisplayedDish>) => {
+        removeDish: (state, action: PayloadAction<DisplayedDish>)  => {
             const dishType = action.payload.dishType;
             const id = action.payload.id;
             if(state.dishes[dishType].includes(id)){